CIV-8291 : Refactor duplicated methods to a common service
